Update 3rd party libs and code cleanup
Updated to the latest version of imGUI (1.79) and native file dialog.
In the process I also fixed all compiler warnings.  Some of the code
has been refactored in order to move most of the SDL code to the ui
module and out of the video area.

+++ nativefiledialog/src/nfd_common.h
@@ -12,6 +12,8 @@
 
 #include "nfd.h"
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -22,13 +24,13 @@ extern "C" {
 #define NFD_UTF8_BOM "\xEF\xBB\xBF"
 
 
-	void  *NFDi_Malloc(size_t bytes);
-	void   NFDi_Free(void *ptr);
-	void   NFDi_SetError(const char *msg);
-	int    NFDi_SafeStrncpy(char *dst, const char *src, size_t maxCopy);
-	size_t NFDi_UTF8_Strlen(const nfdchar_t *str);
-	int    NFDi_IsFilterSegmentChar(char ch);
-
+void  *NFDi_Malloc( size_t bytes );
+void   NFDi_Free( void *ptr );
+void   NFDi_SetError( const char *msg );
+int    NFDi_SafeStrncpy( char *dst, const char *src, size_t maxCopy );
+int32_t NFDi_UTF8_Strlen( const nfdchar_t *str );
+int    NFDi_IsFilterSegmentChar( char ch );
+    
 #ifdef __cplusplus
 }
 #endif
